+++ CHANGELOG.md
@@ -15,7 +15,7 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Add 'either' fault-handling abstraction in `preflex.either` ns (adaptation of the either-monad)
   - Helper fns `success`, `failure`
-  - The `bind` (and `bind-deref->`) dispatcher
+  - The `bind` (and `bind-deref`) dispatcher
   - Helpers `do-either`, `deref-either`
 - Add Hystrix-metrics emulation helpers
   - Command metrics